% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFIG.R
\name{geneOnt}
\alias{geneOnt}
\title{A function for gene ontology enrichment analysis}
\usage{
geneOnt(InputFile = "", IDType = "SYMBOL", OutputName = "")
}
\description{
A function for gene ontology enrichment analysis
}
