% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFIG.R
\name{ontb2bBar}
\alias{ontb2bBar}
\title{Function for visualization of the gene ontology enrichment analysis results}
\usage{
ontb2bBar(
  InputFile = "",
  OntologyCutoff = "10",
  UpBarColor = "lightgreen",
  DownBarColor = "lightblue",
  PointColor = "darkblue"
)
}
\description{
Function for visualization of the gene ontology enrichment analysis results
}
