% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFIG.R
\name{ontBar}
\alias{ontBar}
\title{Function for visualization of the gene ontology enrichment analysis results}
\usage{
ontBar(
  InputFile = "",
  OntologyCutoff = "10",
  BarColor = "lightgreen",
  PointColor = "darkblue"
)
}
\description{
Function for visualization of the gene ontology enrichment analysis results
}
