% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFIG.R
\name{ontBubble}
\alias{ontBubble}
\title{Function for visualization of the gene ontology enrichment analysis results}
\usage{
ontBubble(
  InputFile = "",
  OntologyCutoff = "10",
  BubbleOutline = "darkblue",
  BubbleFill = "lightblue"
)
}
\description{
Function for visualization of the gene ontology enrichment analysis results
}
