% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFIG.R
\name{ontComp}
\alias{ontComp}
\title{Function for gene ontology enrichment analysis comparison}
\usage{
ontComp(InputOne = "", InputTwo = "")
}
\description{
Function for gene ontology enrichment analysis comparison
}
