% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFIG.R
\name{ontVenn}
\alias{ontVenn}
\title{Function for gene ontology enrichment analysis comparison}
\usage{
ontVenn(
  InputFile,
  FirstInputFileName = "Input 1",
  SecondInputFileName = "Input 2",
  FirstInputColor = "lightgreen",
  SecondInputColor = "blue",
  Title = "Gene Ontology Comparison"
)
}
\description{
Function for gene ontology enrichment analysis comparison
}
