% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFIG.R
\name{ontVennSub}
\alias{ontVennSub}
\title{Function for gene ontology enrichment analysis comparison}
\usage{
ontVennSub(
  InputFile,
  FirstInputFileName = "Input 1",
  SecondInputFileName = "Input 2",
  FirstInputColor = "lightgreen",
  SecondInputColor = "blue"
)
}
\description{
Function for gene ontology enrichment analysis comparison
}
