% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{print.gofIG}
\alias{print.gofIG}
\title{Print method for tests of the inverse Gaussian distribution}
\usage{
\method{print}{gofIG}(x, ...)
}
\arguments{
\item{x}{object of class "gofIG".}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
the argument x, invisibly, as for all \link{print} methods.
}
\description{
Printing objects of class "gofIG".
}
\details{
A \code{gofIG} object is a named list of numbers and character string, supplemented with \code{test} (the name of the teststatistic). \code{test} is displayed as a title.
The remaining elements are given in an aligned "name = value" format.
}
\examples{
\donttest{print(test.ABEV1(rgamma(20,1)))}

}
